fix: 鼠标点击应用程序之外的区域，应用程序没有退出
原有退出逻辑只做了窗口状态从激活变为非激活的退出。但是，在某些场景下，鼠标操作不会触发焦点切换，此时激活的窗口仍然是全局搜索。
补充鼠标点击区域判断，在全局搜索外部区域则直接退出。

Log:
Bug: https://pms.uniontech.com/zentao/bug-view-90389.html；https://pms.uniontech.com/zentao/bug-view-90392.html

+++ src/grand-search/business/query/querycontroller.cpp
@@ -78,6 +78,8 @@ void QueryControllerPrivate::onSearchTextChanged(const QString &txt)
 
 void QueryControllerPrivate::onTerminateSearch()
 {
+    if (m_missionId.isEmpty())
+        return;
     qDebug() << "m_daemonDbus->Terminate begin missionId:" << m_missionId;
     m_daemonDbus->Terminate();
     qDebug() << "m_daemonDbus->Terminate end   missionId:" << m_missionId;
